requirements.txt
lightning>=2.2.0
einops>=0.7.0
numpy>=1.24
pillow>=10.0
scikit-image>=0.22
scipy>=1.11
opencv-python>=4.9
tqdm>=4.66
tensorboard>=2.15
matplotlib>=3.8